requirements.txt
annotated-types==0.7.0
anyio==4.4.0
boto3==1.35.2
botocore==1.35.2
click==8.1.7
cloudpickle==3.0.0
copulas==0.11.0
ctgan==0.10.1
deepecho==0.6.0
ecdsa==0.19.0
exceptiongroup==1.2.2
Faker==27.0.0
fastapi==0.112.1
filelock==3.15.4
fsspec==2024.6.1
graphviz==0.20.3
greenlet==3.0.3
h11==0.14.0
idna==3.7
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.20
nvidia-nvtx-cu12==12.1.105
packaging==24.1
pandas==2.2.2
platformdirs==4.2.2
plotly==5.23.0
psycopg2-binary==2.9.9
pyasn1==0.6.0
pydantic==2.8.2
pydantic_core==2.20.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
pytz==2024.1
PyYAML==6.0.2
rdt==1.12.3
rsa==4.9
s3transfer==0.10.2
scikit-learn==1.5.1
scipy==1.14.1
sdmetrics==0.15.1
sdv==1.15.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.32
SQLAlchemy-Utils==0.41.2
starlette==0.38.2
sympy==1.13.2
tenacity==9.0.0
threadpoolctl==3.5.0
torch==2.4.0
tqdm==4.66.5
triton==3.0.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.30.6
